==> reports/03-include_mass.Rmd <==
---
title: "Test whether to include log_mass in phylogenetic linear regression"
author: "Lucas Nell"
date: "`r Sys.setenv(TZ='America/Chicago'); format(Sys.Date(), '%d %b %Y')`"
output:
  github_document:
    toc: true
    toc_depth: 2
editor_options: 
  chunk_output_type: console
---

```{r setup, include = FALSE, cache = FALSE, purl = FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = normalizePath(".."))
# Change this to TRUE if you want to run the models instead of getting them from 
# RDS files:
run_models <- FALSE
```


This file determines whether to use `log_mass` in `phylolm` regressions and 
`cor_phylo` correlations.
These analyses take about half an hour to run.

All p-values below are for whether the coefficient for log(mass) are not equal to zero.


```{r load_packages, echo = FALSE}
suppressPackageStartupMessages({
    library(readr)
    library(dplyr)
    library(tidyr)
    library(purrr)
    library(phylolm)
    library(ape)
    library(phyr)
})
invisible(sapply(list.files('R', '*.R', full.names = TRUE), source))
```
```{r load_models, echo = FALSE, eval = !run_models}
diet_fit <- read_rds('output/inc_mass_diet_fit.rds')
absorp_fit <- read_rds('output/inc_mass_absorp_fit.rds')
spp_fits <- read_rds('output/inc_mass_spp_fits.rds')
pos_fits <- read_rds('output/inc_mass_pos_fits.rds')
clear_sef <- read_rds('output/inc_mass_cor_phylo.rds')
```


# Necessary data:

```{r absorp_data}
# Morphometrics by species
spp_df <- get_df('spp')
tr <- get_tr('spp')
# Morphometrics by species and segment
seg_types <- c('prox', 'mid', 'dist')
pos_df_list <- lapply(seg_types, get_df, .df = 'pos')
names(pos_df_list) <- seg_types
# Absorption by species
absorp_df <- get_df('absorp')
absorp_se_df <- get_df('absorp', .stat = 'se')  # <-- contains standard errors
absorp_tr <- get_tr('absorp')
# Clearance by species
clear_df <- get_df('clear')
clear_se_df <- get_df('clear', .stat = 'se')  # <-- contains standard errors
clear_tr <- get_tr('clear')
```




# `SEF` on `Diet`

`phylolm` call and output:

```{r sef_diet, eval = run_models}
set.seed(29851644)
diet_fit <- phylolm(log_sef ~ diet + log_mass, data = spp_df, phy = tr,
                    model = 'lambda', boot = 2000)
```
```{r write_diet, echo = FALSE, eval = run_models}
write_rds(diet_fit, 'output/inc_mass_diet_fit.rds')
```

P-value:

```{r sef_diet_pval, echo = FALSE}
cat(sprintf('P = %.3f\n', pval(diet_fit, 'log_mass')))
```


# `Absorption` on `Clade`


`phylolm` call and output:

```{r absorp_clade, eval = run_models}
set.seed(1092141389)
absorp_fit <- suppressWarnings(  # gives warning about lambda being very low
    phylolm(log_absorp ~ clade + log_mass, data = absorp_df, phy = absorp_tr, 
            model = 'lambda', boot = 2000)
)
```
```{r write_absorp, echo = FALSE, eval = run_models}
write_rds(absorp_fit, 'output/inc_mass_absorp_fit.rds')
```

P-value:

```{r absorp_clade_pval, echo = FALSE}
cat(sprintf('P = %.3f\n', pval(absorp_fit, 'log_mass')))
```



# `Morphometrics` on `Clade`

`phylolm` calls:

```{r sp_analyses, eval = run_models}
spp_ys <- c("log_intestinal_length", "log_nsa", "log_vill_surface_area",
            "log_total_enterocytes")
set.seed(357885189)
spp_fits <- lapply(
    spp_ys,
    function(y) {
        f <- paste(y, '~ clade + log_mass')
        suppressWarnings(
            do.call("phylolm", list(as.formula(f), data = as.name("spp_df"),
                                    phy = as.name("tr"), model = 'lambda',
                                    boot = 2000))
        )
    })
names(spp_fits) <- spp_ys
```
```{r write_spp, echo = FALSE, eval = run_models}
write_rds(spp_fits, 'output/inc_mass_spp_fits.rds')
```

P-values:

```{r sp_analyses_pvals, echo = FALSE}
cbind(P = sapply(spp_fits, pval, 'log_mass'))
```





# `Morphometrics` on `Clade`, separately by segment


`phylolm` calls:


```{r pos_analyses, eval = run_models}
pos_ys <- c('log_intestinal_diameter', 'log_villus_height', 'villus_width', 
            'crypt_width', 'log_sef', 'enterocyte_diameter', 'log_enterocyte_density')
set.seed(632929430)
pos_fits <- lapply(
    seg_types,
    function(pos) {
        # Assigning to obj named <pos>_df so that the call identifies the position
        assign(paste0(pos, '_df'), pos_df_list[[pos]])
        lapply(
            pos_ys,
            function(y) {
                f <- paste(y, ' ~ clade + log_mass')
                arg_list <- list(
                    formula = as.formula(f),
                    data = as.name(paste0(pos, "_df")),
                    phy = as.name("tr"), model = 'lambda')
                # Some models don't find the peak likelihood unless specifying a
                # starting value of 0.1.
                LL_nostart <- suppressWarnings(do.call("phylolm", arg_list))$logLik
                arg_list_wstart <- c(arg_list, starting.value = quote(list(lambda = 0.1)))
                LL_wstart <- suppressWarnings(do.call("phylolm", arg_list_wstart))$logLik
                if (LL_wstart > LL_nostart) arg_list <- arg_list_wstart
                # Adding number of bootstrap replicates
                arg_list <- c(arg_list, boot = 2000)
                # Now call phylolm
                suppressWarnings(do.call("phylolm", arg_list))
            })
    })
names(pos_fits) <- seg_types
for (i in 1:length(pos_fits)) names(pos_fits[[i]]) <- pos_ys; rm(i)
```
```{r write_pos, echo = FALSE, eval = run_models}
write_rds(pos_fits, 'output/inc_mass_pos_fits.rds')
```


Because there are so many models, I'm writing whether to include mass to a csv file:

```{r csv_pos}
include_df <- lapply(
    seg_types, 
    function(p) {
        tibble(pos = p, y = names(pos_fits[[p]]),
               include = sapply(pos_fits[[p]], pval, 
                                params = 'log_mass') < 0.05)
    }) %>% 
    bind_rows()
write_csv(include_df, 'output/include_mass_pos.csv')
```




P-values:

```{r pos_analyses_pvals, echo = FALSE}
cat("proximal:\n")
cbind(P = sapply(pos_fits$prox, pval, params = 'log_mass'))
cat("middle:\n")
cbind(P = sapply(pos_fits$mid, pval, params = 'log_mass'))
cat("distal:\n")
cbind(P = sapply(pos_fits$dist, pval, params = 'log_mass'))
```








# `Clearance` and `SEF`

In `cor_phylo`, you can include covariate(s) that might be having confounding effects.

Below, I'm trying out whether either clearance or SEF needs body mass as a covariate
by including it separately for clearance, then for SEF.



```{r clear_sef-data}
# Making data frame with all info, including measurement errors:
clear_df_bm <- clear_df
clear_df_bm$log_clear_se <- clear_se_df$log_clear
clear_df_bm$log_sef_se <- clear_se_df$log_sef
# For this comparison, I have to remove one row that doesn't have body mass
clear_df_bm <- clear_df_bm[!is.na(clear_df$log_mass),]

# Phylogenetic tree, removing species with no body mass data
clear_tr_bm <- ape::drop.tip(
    clear_tr,
    tip = clear_tr$tip.label[!clear_tr$tip.label %in% clear_df_bm$species])
```


Now organizing options and running `cor_phylo`.

```{r clear_sef, eval = run_models}
# Options that are the same for both calls
opts <- list(traits = quote(list(log_clear, log_sef)),
             meas_errors = quote(list(log_clear_se, log_sef_se)),
             species = quote(species), data = quote(clear_df_bm),
             phy = quote(clear_tr_bm), boot = 2000,
             method = "nelder-mead-r",
             constrain_d = TRUE,
             max_iter = 1e6, keep_boots = "fail")
# Now separating into different lists:
opts <- list(clear = c(opts, covariates = quote(list(log_clear = log_mass))),
             sef = c(opts, covariates = quote(list(log_sef = log_mass))))

# cor_phylo run with bootstrapping (takes ~1.5 min)
set.seed(1844365955)
clear_sef <- map(c("clear", "sef"), ~ do.call(cor_phylo, opts[[.x]])) %>%
    setNames(c("clear", "sef"))
```

```{r cor_phylo_write, echo = FALSE, eval = run_models}
write_rds(clear_sef, 'output/inc_mass_cor_phylo.rds')
```


Some bootstrap replicates did not converge, so I'm refitting them using a higher
threshold for the reciprocal condition number of two matrices inside the
likelihood function.
This makes the optimization process more strongly "bounce away" from badly
conditioned matrices.
From trial and error, two sets of refits (using `rcond_threshold` values of
`5e-4` and `5e-3`) seem to make all the replicates converge and
provide sensible results.


```{r cor_phylo_refits, eval = TRUE}
set.seed(1110216289)
cp_boot_refits <- list(
    clear = list(
        one = refit_boots(clear_sef$clear, rcond_threshold = 5e-4),
        two = NA
    ),
    sef = list(
        one = refit_boots(clear_sef$sef, rcond_threshold = 5e-4),
        two = NA
    ))
cp_boot_refits$clear$two <- refit_boots(clear_sef$clear,
                                inds = which(map_lgl(cp_boot_refits$clear$one,
                                                     ~ .x$convcode != 0)),
                                rcond_threshold = 5e-3)
cp_boot_refits$sef$two <- refit_boots(clear_sef$sef,
                              inds = which(map_lgl(cp_boot_refits$sef$one,
                                                   ~ .x$convcode != 0)),
                              rcond_threshold = 5e-3)
```



P-values for including body mass for clearance and SEF, respectively:

```{r clear_sef_pval, echo = FALSE}
cat(sprintf('P = %.3f\n',
            pval(clear_sef$clear, params = "B0",
                 refits = cp_boot_refits$clear)["log_clear_log_mass",]))
cat(sprintf('P = %.3f\n',
            pval(clear_sef$sef, params = "B0",
                 refits = cp_boot_refits$sef)["log_sef_log_mass",]))
```




# Session info

This outlines the package versions I used for this script.

```{r session_info, echo = FALSE}
devtools::session_info()
```
