==> reports/04-phylo_fits.Rmd <==
---
title: "Phylogenetic linear regressions and correlations"
author: "Lucas Nell"
date: "`r Sys.setenv(TZ='America/Chicago'); format(Sys.Date(), '%d %b %Y')`"
tz: "CST6CDT"
output:
  github_document:
    toc: true
    toc_depth: 2
editor_options: 
  chunk_output_type: console
---


```{r setup, include = FALSE, cache = FALSE, purl = FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = normalizePath(".."))
run_models <- FALSE
```


This file conducts the linear regressions using `phylolm::phylolm` and 
computes correlations using `phyr::cor_phylo`.



```{r setup2, echo = FALSE, purl = FALSE}
suppressPackageStartupMessages({
    library(readr)
    library(dplyr)
    library(tidyr)
    library(purrr)
    library(phylolm)
    library(ape)
    library(phyr)
})
invisible(sapply(list.files('R', '*.R', full.names = TRUE), source))
```


# Retrieve data

For more information on the functions `get_df` and `get_tr` below (plus `filter_tr` 
and `cp_mat` used later), see [`R/get_data.R`](R/get_data.R).

Note that absorption and clearance data need standard errors as well as means.

```{r get_data}
# Tree for all morphometric and diet analyses
tr <- get_tr('spp')
# Morphometrics by species
spp_df <- get_df('spp')
# Morphometrics by species and segment
seg_types <- c('prox', 'mid', 'dist')
pos_df_list <- lapply(seg_types, get_df, .df = 'pos')
names(pos_df_list) <- seg_types
# Absorption by species
absorp_df <- get_df('absorp')
absorp_se_df <- get_df('absorp', .stat = 'se')
absorp_tr <- get_tr('absorp')
# Clearance by species
clear_df <- get_df('clear')
clear_se_df <- get_df('clear', .stat = 'se')
clear_tr <- get_tr('clear')
```



# `phylolm`

The following sections are regressions using `phylolm::phylolm`.

## `SEF` on `Diet`

`phylolm` call:

```{r sef_diet, eval = run_models}
set.seed(581120)
diet_fit <- phylolm(log_sef ~ diet, data = spp_df, phy = tr,
                    model = 'lambda', boot = 2000)
```

Saving output:

```{r save_diet, eval = run_models, purl = FALSE}
readr::write_rds(diet_fit, 'output/models_diet.rds')
```
```{r sef_diet_load, echo = FALSE, purl = FALSE, eval = !run_models}
diet_fit <- readr::read_rds('output/models_diet.rds')
```



## `Absorption` on `Clade`

"Absorption" here means `Fractional absorption / (total intestinal surface)`,
where `total intestinal surface = NSA * SEF`

`phylolm` call:

```{r absorp_clade, eval = run_models}
set.seed(454094511)
absorp_fit <- suppressWarnings(  # gives warning about lambda being very low
    phylolm(log_absorp ~ clade + log_mass, data = absorp_df, phy = absorp_tr, 
            model = 'lambda', boot = 2000)
)
```


Saving output:

```{r save_absorp, eval = run_models, purl = FALSE}
readr::write_rds(absorp_fit, 'output/models_absorp.rds')
```

```{r absorp_load, echo = FALSE, purl = FALSE, eval = !run_models}
absorp_fit <- readr::read_rds('output/models_absorp.rds')
```




## `Morphometrics` on `Clade`

List of `Morphometrics`:

- Intestinal length
- NSA
- Villus surface area
- Total number of enterocytes (log-transformed)
    * Calculated as such: `log(NSA * enterocyte_density)`

> log(body mass) as covariate for all

These are the column names for the above parameters:
```{r sp_analyses_cols}
spp_ys <- c("log_intestinal_length", "log_nsa", "log_vill_surface_area",
            "log_total_enterocytes")
```


`phylolm` calls:


```{r sp_analyses, eval = run_models}
# (takes ~7.5 min)
set.seed(88754829)
spp_fits <- lapply(
    spp_ys,
    function(y_) {
        f <- paste(y_, '~ clade + log_mass')
        suppressWarnings(
            do.call("phylolm", list(as.formula(f), data = as.name("spp_df"),
                                    phy = as.name("tr"), model = 'lambda',
                                    boot = 2000))
        )
    })
names(spp_fits) <- spp_ys
```

Saving output:

```{r save_spp, eval = run_models, purl = FALSE}
readr::write_rds(spp_fits, 'output/models_spp.rds')
```
```{r spp_load, echo = FALSE, purl = FALSE, eval = !run_models}
spp_fits <- readr::read_rds('output/models_spp.rds')
```




## `Morphometrics` on `Clade`, separately by segment


(Segment = proximal, middle, or distal)

List of `Y`s:

- Intestinal diameter (log-transformed, body mass as covariate)
- Villus height (body mass as covariate)
- Villus width
- Crypt width
- Surface enlargement factor (SEF)
- Enterocyte diameter
- Enterocytes per cm^2 NSA (log-transformed)


Below are the column names for these parameters and all the segment types.

```{r pos_ys}
pos_ys <- c('log_intestinal_diameter', 'log_villus_height', 'villus_width', 
            'crypt_width', 'log_sef', 'enterocyte_diameter', 'log_enterocyte_density')
seg_types <- c('prox', 'mid', 'dist')
```


Below is a data frame including whether or not to include `log_mass` as a covariate.
This determination was based on whether `log_mass` had a significant effect when it was
included in the model, where p-values were based on parametric bootstrapping 
(see [`docs/03-include_mass`](docs/03-include_mass.md)).

```{r pos_log_mass_include}
include_mass <- read_csv('output/include_mass_pos.csv', col_types = 'ccl')
include_mass
```


`phylolm` call:

```{r pos_analyses, eval = run_models}
# (takes ~32 min)
set.seed(25413535)
pos_fits <- lapply(
    seg_types,
    function(pos_) {
        # Assigning to obj named <pos_>_df so that the call identifies the position
        assign(paste0(pos_, '_df'), get_df('pos', .pos = pos_))
        lapply(
            pos_ys,
            function(y_) {
                f <- paste(y_, ' ~ clade')
                # Whether to include log_mass covariate
                imc <- {include_mass %>% filter(pos == pos_, y == y_)}$include
                if (imc[1]) f <- paste(f, '+ log_mass')
                arg_list <- list(
                    as.formula(f),
                    data = as.name(paste0(pos_, "_df")),
                    phy = as.name("tr"), model = 'lambda')
                # Some models don't find the peak likelihood unless specifying a
                # starting value of 0.1.
                LL_nostart <- suppressWarnings(do.call("phylolm", arg_list))$logLik
                arg_list_wstart <- c(arg_list, starting.value = quote(list(lambda = 0.1)))
                LL_wstart <- suppressWarnings(do.call("phylolm", arg_list_wstart))$logLik
                if (LL_wstart > LL_nostart) arg_list <- arg_list_wstart
                arg_list <- c(arg_list, boot = 2000)
                # Now create the final phylolm object
                suppressWarnings(do.call("phylolm", arg_list))
            })
    })
names(pos_fits) <- seg_types
for (i in 1:length(pos_fits)) names(pos_fits[[i]]) <- pos_ys; rm(i)
```


The model for `crypt_width ~ clade` in the proximal segment has a higher log likelihood
at a very high phylogenetic signal (`logLik = 68.14` at `lambda = 0.993`) than at 
very low signal (`logLik = 66.88` at `lambda = 1e-7`).
However, this model is sensitive to starting values, which suggests multiple peaks in
the likelihood profile.
Moreover, the models for the other segments show very low phylogenetic signal 
(`1e-7` for both), and this model re-run with the Ornstein-Uhlenbeck model for
phylogenetic error ("OU"; `OUfixedRoot` in `phylolm`) has a higher log likelihood and
shows a much lower phylogenetic signal (`logLik = 68.54` and `alpha = 0.0156`).
Thus the model likely had convergence issues using Pagel's lambda, so I'm 
replacing the original model with one using the OU error model below.
I'm saving the original one to report it, too.

```{r pos_analyses_crypt_width_prox, eval = run_models}
pos_fits$prox$crypt_width_pagel <- pos_fits$prox$crypt_width
prox_df <- get_df('pos', .pos = 'prox')
set.seed(1340481016)
pos_fits$prox$crypt_width <- update(pos_fits$prox$crypt_width, 
                                       model = "OUfixedRoot")
```


Saving output:

```{r save_pos, eval = run_models, purl = FALSE}
readr::write_rds(pos_fits, 'output/models_pos.rds')
```
```{r pos_load, echo = FALSE, purl = FALSE, eval = !run_models}
pos_fits <- readr::read_rds('output/models_pos.rds')
```












# `cor_phylo`

From the original manuscript: 

> ... we used reduced major axis regression (model II regression)... because both 
> variables [X and Y] were subject to error

Instead of an RMA regression, I'll be using `phyr::cor_phylo`, which is similar to
`ape::corphylo` but faster, more accurate, and capable of conducting parametric
bootstrapping.
P-values are calculated using bootstrap replicates.

I used the same P-values to determine that I do not need to use `log_mass` as
an independent variable for any of these fits
(see [`docs/03-include_mass`](docs/03-include_mass.md) for more info).

All variables under this section are log-transformed.


```{r cor_phylo_load, echo = FALSE, eval = !run_models}
# Load all cor_phylo objects
clear_sef <- readr::read_rds('output/models_cor_phylo.rds')
cp_boot_refits <- readr::read_rds('output/models_cor_phylo_refits.rds')
```


## `Clearance` and `SEF`

Clearance = "paracellular probe L-arabinose clearance"


```{r clear_sef, eval = run_models}
# Making data frame with both means and SEs:
clear_sef_df <- bind_cols(clear_df %>% select(species, log_sef, log_clear),
                          clear_se_df %>% select(log_sef, log_clear) %>% 
                              rename(log_sef_se = log_sef, log_clear_se = log_clear))
set.seed(1844365955)
clear_sef <- cor_phylo(traits = list(log_sef, log_clear),
                       meas_errors = list(log_sef_se, log_clear_se),
                       species = species, phy = clear_tr, data = clear_sef_df,
                       method = "nelder-mead-r", constrain_d = TRUE,
                       boot = 2000, max_iter = 1e6)
```


Some bootstrap replicates did not converge, so I'm refitting them using a higher
threshold for the reciprocal condition number of two matrices inside the
likelihood function.
This makes the optimization process more strongly "bounce away" from badly
conditioned matrices.
From trial and error, two sets of refits (using `rcond_threshold` values of
`1e-4` and `2e-3`) seem to make all the replicates converge and
provide sensible results.


```{r cor_phylo_refits, eval = run_models}
cp_boot_refits <- list(
        one = refit_boots(clear_sef, rcond_threshold = 1e-4),
        two = NA
    )
cp_boot_refits$two <- refit_boots(clear_sef,
                                inds = which(map_lgl(cp_boot_refits$one,
                                                     ~ .x$convcode != 0)),
                                rcond_threshold = 2e-3)
```


Saving `cor_phylo` and refits (class `cp_refits`) objects:

```{r save_cor_phylo, eval = run_models}
readr::write_rds(clear_sef, 'output/models_cor_phylo.rds')
readr::write_rds(cp_boot_refits, 'output/models_cor_phylo_refits.rds')
```





# Assembling all output into one table


I ran `summ_df` on all models (both `phylolm` and `cor_phylo`) above.
This function summarizes both of these object classes into a single data frame.
See [`R/model_summaries.R`](R/model_summaries.R) for more info.

```{r mod_summaries, purl = FALSE}
mod_summaries <- bind_rows(
    list(
        summ_df(diet_fit),
        summ_df(absorp_fit),
        bind_rows(lapply(spp_fits, summ_df)),
        bind_rows(
            lapply(names(pos_fits), function(p) {
                bind_rows(lapply(pos_fits[[p]], summ_df, .pos = p))
            })),
        summ_df(clear_sef, cp_boot_refits)
        ))
```


I lastly write this summary to a csv file.


```{r write_summaries, purl = FALSE}
write_csv(mod_summaries, 'output/models_summaries.csv')
```




# Session info

This outlines the package versions I used for these analyses.

```{r session_info, echo = FALSE, purl = FALSE}
devtools::session_info()
```
